apps/ops-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musematic-ops-cli"
version = "0.1.0"
description = "Standalone installer and operations CLI for the Musematic platform."
requires-python = ">=3.12"
dependencies = [
  "typer[all]>=0.12",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "httpx>=0.27",
  "asyncpg>=0.29",
  "aiokafka>=0.12",
  "redis>=5.0",
  "neo4j>=5.22",
  "clickhouse-connect>=0.8",
  "grpcio>=1.65",
  "grpcio-health-checking>=1.65",
  "cryptography>=43.0",
  "aioboto3>=13.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.7",
  "mypy>=1.11",
  "pyinstaller>=6.0",
]

[project.scripts]
platform-cli = "platform_cli.main:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src", "tests"]
testpaths = ["tests"]
addopts = "-ra"
markers = [
  "integration: requires subprocesses or external services",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "N",
  "UP",
  "B",
  "A",
  "ASYNC",
  "C4",
  "DTZ",
  "G",
  "LOG",
  "PIE",
  "PT",
  "PTH",
  "RUF",
]
ignore = [
  "B008",
  "ASYNC220",
  "ASYNC221",
  "ASYNC240",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
plugins = ["pydantic.mypy"]
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = [
  "aioboto3",
  "aiokafka",
  "aiokafka.admin",
  "asyncpg",
  "clickhouse_connect",
  "grpc",
  "grpc.*",
  "grpc_health",
  "grpc_health.*",
  "neo4j",
  "typer",
  "redis",
  "redis.*",
  "yaml",
]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["src/platform_cli"]

[tool.coverage.report]
skip_covered = false
show_missing = true
fail_under = 95
